Convert to async stack and move DI wiring out of business classes
- Async throughout: httpx.AsyncClient for SWAPI, SQLAlchemy AsyncSession
  over asyncpg, await through router -> service -> client. Pure logic
  (domain, utils) stays sync.
- Services and clients are now plain Python: all Depends wiring lives in
  app/dependencies.py as provider functions.
- Shared httpx client created in lifespan with base_url and timeout;
  HTTP errors wrapped into app exceptions at the client boundary.
- Transaction boundary moved from database clients into get_db_session:
  commit on success, rollback on error; clients only add/flush.
- Fixes: vehicle efficiency was computed but never persisted; efficiency
  column Integer -> Float (with migration); GET missing vehicle now 404s
  instead of 500ing.
- Tests: pytest-asyncio (auto mode), httpx.MockTransport replaces
  responses, AsyncMock-based session mocks replace mock-alchemy.

+++ app/models/star_wars_vehicle_model.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy import Integer, String
+from sqlalchemy import Float, Integer, String
 from database import Base
 
 
@@ -19,4 +19,4 @@ class StarWarsVehicle(Base):
     consumables: Mapped[str] = mapped_column(String, nullable=True)
     vehicle_class: Mapped[str] = mapped_column(String, nullable=True)
     pilots: Mapped[str] = mapped_column(String, nullable=True)
-    efficiency: Mapped[float] = mapped_column(Integer, nullable=True)
+    efficiency: Mapped[float] = mapped_column(Float, nullable=True)
